third_party/blink/web_tests/xr/events_referenceSpace_reset.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="../external/wpt/resources/chromium/webxr-test.js"></script>
<script src="../external/wpt/webxr/resources/webxr_test_constants.js"></script>
<script src="../xr/resources/xr-internal-device-mocking.js"></script>
<script src="../xr/resources/xr-test-utils.js"></script>
<canvas id="webgl-canvas"></canvas>

<script>
let testName =
  "XRSession resetpose from a device properly fires off the right events"

let watcherDone = new Event("watcherdone");

let fakeDeviceInitParams = TRACKED_IMMERSIVE_DEVICE;

let requestSessionModes = [
  'inline',
  'immersive-vr',
];

let testFunction = function(session, t, fakeDeviceController) {
  // TODO(981003): Because the device gets re-used we have to ensure that the
  // resetPose bit is cleared.
  fakeDeviceController.setResetPose(false);
  // Session must have a baseLayer or frame requests will be ignored.
  session.updateRenderState({
    baseLayer: new XRWebGLLayer(session, gl, {
        compositionDisabled: session.mode == 'inline' })
  });

  let resetPromise = session.requestReferenceSpace('local')
  .then((refSpace) => {
    let eventWatcher = new EventWatcher(
        t, refSpace, ["reset", "watcherdone"]);
    refSpace.addEventListener("reset", (event) => {
      assert_equals(event.referenceSpace, refSpace);
      refSpace.dispatchEvent(watcherDone);
    }, false);
    return eventWatcher.wait_for(["reset", "watcherdone"]);
  });

  fakeDeviceController.setResetPose(true);

  // Trigger the reset pose.
  session.requestAnimationFrame(() => {
    session.requestAnimationFrame(() => {});
  });

  return resetPromise;
};

xr_session_promise_test(
  testFunction, fakeDeviceInitParams, requestSessionModes, testName);

</script>
